backend/pyproject.toml
[project]
name = "swesphere-backend"
version = "1.0.0"
description = "FastAPI backend for SWESphere - a Twitter-like social media platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "SWESphere Team" }]
keywords = ["fastapi", "social-media", "api", "async", "postgresql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]

dependencies = [
    # ===================
    # FastAPI & Server
    # ===================
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "orjson>=3.9.10",
    
    # ===================
    # Database
    # ===================
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "greenlet>=3.0.0",
    
    # ===================
    # Redis & Caching
    # ===================
    "redis>=5.0.1",
    
    # ===================
    # Authentication & Security
    # ===================
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.1.2",
    
    # ===================
    # Validation & Settings
    # ===================
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    
    # ===================
    # Background Tasks
    # ===================
    "celery[redis]>=5.3.6",
    
    # ===================
    # HTTP Client
    # ===================
    "httpx>=0.26.0",
    
    # ===================
    # Utils
    # ===================
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "httpx>=0.26.0",
    "aiosqlite>=0.19.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    
    # Code Quality
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    
    # Type stubs
    "types-redis>=4.6.0",
    "types-passlib>=1.7.7",
    "types-python-jose>=3.3.4",
]

docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
swesphere = "app.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ===================
# Ruff Configuration
# ===================
[tool.ruff]
target-version = "py311"
line-length = 88
fix = true

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
    "ARG",   # flake8-unused-arguments
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "PTH",   # flake8-use-pathlib
    "ERA",   # eradicate (commented code)
    "RUF",   # Ruff-specific rules
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function calls in argument defaults
    "C901",   # Too complex
    "ARG001", # Unused function argument (FastAPI deps)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

# ===================
# MyPy Configuration
# ===================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "celery.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "passlib.*"
ignore_missing_imports = true

# ===================
# Pytest Configuration
# ===================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "-ra",
    "--strict-markers",
    "--strict-config",
    "-p", "no:warnings",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# ===================
# Coverage Configuration
# ===================
[tool.coverage.run]
source = ["app"]
branch = true
omit = [
    "app/workers/*",
    "*/migrations/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
show_missing = true
skip_covered = true
fail_under = 80
